refs #3827
Fixed the issued related to ModifyCardAlias of NFC Card. Previously, if no cardPAN is given, list of all pockets are returned as part of the query and first pocket among the list is picked up and updated which is a bug. Added the validation checks to not allow null value against the cardPAN.