Merged revision(s) 991 from branches/MFS_Uangku_Simobi: Modified to fix the SMS issue in Approve Reject subscriber in Simobi.
Issue: We are sending two SMS in this case and in the values to SMS layer are sending using the class variables. So the async method is over riding the old values with latest values so that same sms is delivering twice.

fix: replaced the class varibales in BSIMSMSServiceImpl.java with method variables.